refactor(lib): drop dead utilities, rename media-transcribe → mime
- Delete async-handlers.ts and safe-async.ts (zero imports repo-wide).
- Delete old mime.ts (detectMimeType, zero imports) and the dead
  IMAGE/AUDIO/VIDEO/DOCUMENT_MIMES const objects + type aliases.
- Rename media-transcribe.ts → mime.ts (it holds MIME↔ext maps, no
  transcription) and move normalizeApiBaseUrl into its only caller,
  the OpenAI media provider.
- Dedupe MEDIA_TYPE_LABELS: export once from base-adapter, import in
  the whatsapp adapter.

Net -198 LOC. No behavior change; MIME_TO_AUDIO_EXT misuse in openai.ts
left untouched and flagged for a separate fix.
